app theme as "styled"
