+++ src/index.tsx
@@ -3,10 +3,8 @@ import { createRoot } from 'react-dom/client'
 import SwiperCore, { Autoplay, EffectFade, Navigation, Pagination, Thumbs } from 'swiper'
 import { ApolloProvider } from '@apollo/client'
 import { Router } from 'react-router-dom'
-import { ThemeProvider, CssBaseline } from '@material-ui/core'
-import { ThemeProvider as StyThemeProvider } from 'styled-components'
-import { darkTheme, lightTheme } from './utils/theme'
-import { darkTheme as StDark, lightTheme as StLight } from './utils/st_theme'
+import { ThemeProvider } from 'styled-components'
+import { darkTheme, lightTheme, GlobalStyles } from './utils/theme'
 import App from './App'
 import { useTheme } from './hooks'
 import history from './utils/history'
@@ -29,14 +27,12 @@ const Application: React.FC = () => {
   return (
     <Router history={history}>
       <ThemeProvider theme={theme === 'light' ? lightTheme : darkTheme}>
-        <StyThemeProvider theme={theme === 'light' ? StLight : StDark}>
-          <ModalProvider>
-            <ApolloProvider client={client}>
-              <CssBaseline />
-              <App themeChanger={changeTheme} />
-            </ApolloProvider>
-          </ModalProvider>
-        </StyThemeProvider>
+        <ModalProvider>
+          <ApolloProvider client={client}>
+            <GlobalStyles />
+            <App themeChanger={changeTheme} />
+          </ApolloProvider>
+        </ModalProvider>
       </ThemeProvider>
     </Router>
   )
