+++ src/utils/theme.ts
@@ -1,126 +1,116 @@
-import { createTheme } from '@material-ui/core/styles'
-import { BreakpointOverrides } from '@material-ui/core/styles/createBreakpoints'
+import { DefaultTheme, createGlobalStyle } from 'styled-components'
 
-// #ff9900 - orange secondary
+export enum ThemeType {
+  light = 'light',
+  dark = 'dark'
+}
 
-declare module '@material-ui/core/styles/createBreakpoints' {
-  interface BreakpointOverrides {
-    xs: true
-    sm: true
-    md: true
-    lg: true
-    xl: true
-    tablet: true
-    laptop: true
-    desktop: true
+interface Breakpoints {
+  xs: number
+  sm: number
+  md: number
+  lg: number
+  xl: number
+  tablet: number
+  laptop: number
+  desktop: number
+}
+
+export type ITheme = {
+  type: ThemeType
+  background: string
+  colors: {
+    primary: string
+    main: string
+    light: string
+    dark: string
+  }
+  media: {
+    xs: string
+    sm: string
+    md: string
+    lg: string
+    xl: string
+    tablet: string
+    laptop: string
+    desktop: string
   }
 }
 
-declare module '@material-ui/core/styles/createTheme' {
-  interface Theme {
-    appDrawer: {
-      width: React.CSSProperties['width']
-      breakpoint: BreakpointOverrides
-    }
-  }
-  // allow configuration using `createTheme`
-  interface ThemeOptions {
-    appDrawer?: {
-      width?: React.CSSProperties['width']
-      breakpoint?: BreakpointOverrides
-    }
-  }
-}
-
-const FONTS = [
-  'Montserrat',
-  'Roboto',
-  'Arial',
-  'sans-serif',
-  '"Apple Color Emoji"',
-  '"Segoe UI Emoji"',
-  '"Segoe UI Symbol"'
-].join(',')
+// **********************************
 
 const BRAND_COLOR = '#F2E30C'
 
-export const darkTheme = createTheme({
-  breakpoints: {
-    keys: ['xs', 'sm', 'md', 'lg', 'xl', 'tablet', 'laptop', 'desktop'],
-    values: {
-      xs: 0,
-      sm: 450,
-      md: 600,
-      lg: 900,
-      xl: 1200,
-      tablet: 800,
-      laptop: 1000,
-      desktop: 1400
-    }
-  },
-  palette: {
-    type: 'dark',
-    primary: {
-      main: '#ff9900',
-      light: '#fff',
-      dark: '#303030'
-    },
-    secondary: {
-      main: BRAND_COLOR
-    },
-    background: {
-      default: '#1e1e1e'
-    }
-  },
-  typography: {
-    fontFamily: FONTS
-  },
-  overrides: {
-    MuiBackdrop: {
-      root: {
-        backdropFilter: 'blur(3px)',
-        backgroundColor: 'rgba(0, 0, 0, 0.4)'
-      }
-    }
-  }
-})
+const breakpoints: Breakpoints = {
+  xs: 0,
+  sm: 450,
+  md: 600,
+  lg: 900,
+  xl: 1200,
+  tablet: 800,
+  laptop: 1000,
+  desktop: 1400
+}
 
-export const lightTheme = createTheme({
-  breakpoints: {
-    keys: ['xs', 'sm', 'md', 'lg', 'xl', 'tablet', 'laptop', 'desktop'],
-    values: {
-      xs: 0,
-      sm: 450,
-      md: 600,
-      lg: 900,
-      xl: 1200,
-      tablet: 800,
-      laptop: 1000,
-      desktop: 1400
-    }
+const media = {
+  xs: `(min-width: ${breakpoints.xs}px)`,
+  sm: `(min-width: ${breakpoints.sm}px)`,
+  md: `(min-width: ${breakpoints.md}px)`,
+  lg: `(min-width: ${breakpoints.lg}px)`,
+  xl: `(min-width: ${breakpoints.xl}px)`,
+  tablet: `(min-width: ${breakpoints.tablet}px)`,
+  laptop: `(min-width: ${breakpoints.laptop}px)`,
+  desktop: `(min-width: ${breakpoints.desktop}px)`
+}
+
+export const lightTheme: DefaultTheme = {
+  type: ThemeType.light,
+  background: '#fff',
+  colors: {
+    primary: BRAND_COLOR,
+    main: '#383838',
+    light: '#fff',
+    dark: '#303030' // TODO: Same?
   },
-  palette: {
-    type: 'light',
-    primary: {
-      main: '#383838',
-      light: '#fff'
-    },
-    secondary: {
-      main: BRAND_COLOR
-    },
-    background: {
-      default: '#fafafa'
-    }
+  media
+}
+
+export const darkTheme: DefaultTheme = {
+  type: ThemeType.dark,
+  colors: {
+    primary: BRAND_COLOR,
+    main: '#ff9900',
+    light: '#fff',
+    dark: '#303030'
   },
-  typography: {
-    fontFamily: FONTS
-  },
-  overrides: {
-    MuiBackdrop: {
-      root: {
-        backdropFilter: 'blur(3px)',
-        backgroundColor: 'rgba(0, 0, 0, 0.3)'
-      }
-    }
+  background: '#1e1e1e',
+  media
+}
+
+export const GlobalStyles = createGlobalStyle`
+  *,
+  *:after,
+  *:before {
+    -webkit-box-sizing: border-box;
+    -moz-box-sizing: border-box;
+    box-sizing: border-box;
   }
-})
+  body {
+    margin: 0;
+    font-family: Montserrat,Roboto,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol";
+  }
+  input::-webkit-outer-spin-button,
+  input::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+  }
+  input[type=number] {
+    -moz-appearance: textfield;
+  }
+ 
+  /* ul {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+  } */
+`
